:recycle: 責務のリファクタ
